+++ server/src/.mockery.yaml
@@ -49,6 +49,16 @@ packages:
       HealthDatabase:
         all: true
 
+  scrumlr.io/server/notes:
+    config:
+      dir: "{{.InterfaceDir}}"
+      inpackage: true
+    interfaces:
+      NotesService:
+        all: true
+      NotesDatabase:
+        all: true
+
   scrumlr.io/server/sessions:
     config:
       dir: "{{.InterfaceDir}}"
