+++ server/src/database/votings.go
@@ -3,12 +3,13 @@ package database
 import (
 	"context"
 	"errors"
+	"fmt"
 	"github.com/google/uuid"
 	"github.com/uptrace/bun"
 	"scrumlr.io/server/common"
 	"scrumlr.io/server/common/filter"
 	"scrumlr.io/server/database/types"
-
+	"scrumlr.io/server/notes"
 	"time"
 )
 
@@ -105,25 +106,24 @@ func (d *Database) UpdateVoting(update VotingUpdate) (Voting, error) {
 }
 
 func (d *Database) getRankUpdateQueryForClosedVoting(votingQuery string) *bun.UpdateQuery {
-	//newRankSelect := d.db.NewSelect().
-	//	TableExpr("notes as note").
-	//	ColumnExpr(fmt.Sprintf(
-	//		"ROW_NUMBER() OVER (PARTITION BY \"column\" ORDER BY "+
-	//			"(SELECT COUNT(*) FROM notes AS n INNER JOIN (SELECT * FROM VOTES WHERE voting = (SELECT id FROM \"%s\")) as v ON n.id = v.note WHERE n.id = note.id OR n.stack = note.id), rank)-1 AS new_rank",
-	//		votingQuery)).
-	//	Column("id").
-	//	Where(fmt.Sprintf("stack IS NULL AND board = (SELECT board FROM \"%s\")", votingQuery)).
-	//	GroupExpr("id")
-	//
-	//rankUpdate := d.db.NewUpdate().With("_data", newRankSelect).
-	//	Model((*notes.NoteDB)(nil)).
-	//	TableExpr("_data").
-	//	Set("rank = _data.new_rank").
-	//	WhereOr("note.id = _data.id").
-	//	WhereOr("note.stack = _data.id")
-	//
-	//return rankUpdate
-	return nil
+	newRankSelect := d.db.NewSelect().
+		TableExpr("notes as note").
+		ColumnExpr(fmt.Sprintf(
+			"ROW_NUMBER() OVER (PARTITION BY \"column\" ORDER BY "+
+				"(SELECT COUNT(*) FROM notes AS n INNER JOIN (SELECT * FROM VOTES WHERE voting = (SELECT id FROM \"%s\")) as v ON n.id = v.note WHERE n.id = note.id OR n.stack = note.id), rank)-1 AS new_rank",
+			votingQuery)).
+		Column("id").
+		Where(fmt.Sprintf("stack IS NULL AND board = (SELECT board FROM \"%s\")", votingQuery)).
+		GroupExpr("id")
+
+	rankUpdate := d.db.NewUpdate().With("_data", newRankSelect).
+		Model((*notes.NoteDB)(nil)).
+		TableExpr("_data").
+		Set("rank = _data.new_rank").
+		WhereOr("note.id = _data.id").
+		WhereOr("note.stack = _data.id")
+
+	return rankUpdate
 }
 
 func (d *Database) GetVoting(board, id uuid.UUID) (Voting, []Vote, error) {
