+++ server/src/services/services.go
@@ -57,6 +57,17 @@ type Votings interface {
 	GetVotes(ctx context.Context, f filter.VoteFilter) ([]*dto.Vote, error)
 }
 
+type Votings interface {
+	Create(ctx context.Context, body votes.VotingCreateRequest) (*votes.Voting, error)
+	Update(ctx context.Context, body votes.VotingUpdateRequest) (*votes.Voting, error)
+	Get(ctx context.Context, board, id uuid.UUID) (*votes.Voting, error)
+	List(ctx context.Context, board uuid.UUID) ([]*votes.Voting, error)
+
+	AddVote(ctx context.Context, req dto.VoteRequest) (*dto.Vote, error)
+	RemoveVote(ctx context.Context, req dto.VoteRequest) error
+	GetVotes(ctx context.Context, f filter.VoteFilter) ([]*dto.Vote, error)
+}
+
 type BoardReactions interface {
 	Create(ctx context.Context, board uuid.UUID, body dto.BoardReactionCreateRequest)
 }
