+++ server/src/votes/dto.go
@@ -3,6 +3,7 @@ package votes
 import (
 	"github.com/google/uuid"
 	"scrumlr.io/server/database/types"
+	"scrumlr.io/server/notes"
 )
 
 // VotingCreateRequest represents the request to create a new voting session.
@@ -46,18 +47,6 @@ type VotingResultsPerNote struct {
 }
 
 type VotingUpdated struct {
-	//Notes  notes.NoteSlice `json:"notes"`
-	Voting *Voting `json:"voting"`
-}
-
-func Votings(votings []VotingDB, votes []VoteDB) []*Voting {
-	if votings == nil {
-		return nil
-	}
-
-	list := make([]*Voting, len(votings))
-	for index, voting := range votings {
-		list[index] = new(Voting).From(voting, votes)
-	}
-	return list
+	Notes  notes.NoteSlice `json:"notes"`
+	Voting *Voting         `json:"voting"`
 }
