+++ server/src/votes/service.go
@@ -1,195 +1,143 @@
 package votes
 
 import (
-	"context"
-	"database/sql"
-	"errors"
 	"github.com/google/uuid"
-	"scrumlr.io/server/common"
-	"scrumlr.io/server/common/filter"
+	"net/http"
 	"scrumlr.io/server/database"
 	"scrumlr.io/server/database/types"
-	"scrumlr.io/server/logger"
-	notes2 "scrumlr.io/server/notes"
-	"scrumlr.io/server/realtime"
+	"scrumlr.io/server/notes"
+	"scrumlr.io/server/technical_helper"
 )
 
-type Service struct {
-	database VotingDatabase
-	realtime *realtime.Broker
+func (v *Voting) From(voting database.Voting, votes []database.Vote) *Voting {
+	v.ID = voting.ID
+	v.VoteLimit = voting.VoteLimit
+	v.AllowMultipleVotes = voting.AllowMultipleVotes
+	v.ShowVotesOfOthers = voting.ShowVotesOfOthers
+	v.Status = voting.Status
+	v.VotingResults = getVotingWithResults(voting, votes)
+	return v
 }
 
-func (s Service) Create(ctx context.Context, body VotingCreateRequest) (*Voting, error) {
-	log := logger.FromContext(ctx)
-	voting, err := s.database.CreateVoting(VotingInsert{
-		Board:              body.Board,
-		VoteLimit:          body.VoteLimit,
-		AllowMultipleVotes: body.AllowMultipleVotes,
-		ShowVotesOfOthers:  body.ShowVotesOfOthers,
-		Status:             types.VotingStatusOpen,
-	})
-
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, common.BadRequestError(errors.New("only one open voting session is allowed"))
-		}
-		log.Errorw("unable to create voting", "board", body.Board, "error", err)
-		return nil, common.InternalServerError
-	}
-	s.createdVoting(body.Board, voting.ID)
-
-	return new(Voting).From(voting, nil), err
+func (*Voting) Render(_ http.ResponseWriter, _ *http.Request) error {
+	return nil
 }
 
-func (s Service) Update(ctx context.Context, body VotingUpdateRequest) (*Voting, error) {
-	log := logger.FromContext(ctx)
-	if body.Status == types.VotingStatusOpen {
-		return nil, common.BadRequestError(errors.New("not allowed ot change to open state"))
+func Votings(votings []database.Voting, votes []database.Vote) []*Voting {
+	if votings == nil {
+		return nil
 	}
 
-	voting, err := s.database.UpdateVoting(VotingUpdate{
-		ID:     body.ID,
-		Board:  body.Board,
-		Status: body.Status,
-	})
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, common.NotFoundError
-		}
-		log.Errorw("unable to update voting", "err", err)
-		return nil, common.InternalServerError
+	list := make([]*Voting, len(votings))
+	for index, voting := range votings {
+		list[index] = new(Voting).From(voting, votes)
 	}
-
-	if voting.Status == types.VotingStatusClosed {
-		receivedVotes, err := s.getVotes(ctx, body.Board, body.ID)
-		if err != nil {
-			log.Errorw("unable to get votes", "err", err)
-			return nil, err
-		}
-		s.updatedVoting(body.Board, voting.ID)
-		return new(Voting).From(voting, receivedVotes), err
-	}
-	s.updatedVoting(body.Board, voting.ID)
-	return new(Voting).From(voting, nil), err
+	return list
 }
 
-func (s Service) Get(ctx context.Context, board, id uuid.UUID) (*Voting, error) {
-	log := logger.FromContext(ctx)
-	voting, _, err := s.database.GetVoting(board, id)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, common.NotFoundError
+func (v *Voting) UpdateVoting(notes notes.NoteSlice) *VotingUpdated {
+	if v.hasNoResults() {
+		return &VotingUpdated{
+			Notes:  notes,
+			Voting: v,
 		}
-		log.Errorw("unable to get voting session", "voting", id, "error", err)
-		return nil, common.InternalServerError
 	}
 
-	if voting.Status == types.VotingStatusClosed {
-		receivedVotes, err := s.getVotes(ctx, board, id)
-		if err != nil {
-			log.Errorw("unable to get votes", "voting", id, "error", err)
-			return nil, err
-		}
-		return new(Voting).From(voting, receivedVotes), err
+	v.VotingResults = v.calculateTotalVoteCount(notes)
+
+	return &VotingUpdated{
+		Notes:  notes,
+		Voting: v,
 	}
-	return new(Voting).From(voting, nil), err
 }
 
-func (s Service) List(ctx context.Context, board uuid.UUID) ([]*Voting, error) {
-	log := logger.FromContext(ctx)
-	votings, receivedVotes, err := s.database.GetVotings(board)
+func UnmarshallVoteData(data interface{}) (*VotingUpdated, error) {
+	vote, err := technical_helper.Unmarshal[VotingUpdated](data)
+
 	if err != nil {
-		log.Errorw("unable to get votings", "board", board, "error", err)
 		return nil, err
 	}
-	return VotingWithVotes(votings, receivedVotes), err
+
+	return vote, nil
 }
 
-func (s Service) AddVote(ctx context.Context, body VoteRequest) (*Vote, error) {
-	log := logger.FromContext(ctx)
-	v, err := s.database.AddVote(body.Board, body.User, body.Note)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, common.ForbiddenError(errors.New("voting limit reached or no active voting session found"))
-		}
-		log.Warnw("unable to add vote", "board", body.Board, "user", body.User, "note", body.Note, "err", err)
-		return nil, err
-	}
-	return new(Vote).From(v), err
-}
-
-func (s Service) RemoveVote(ctx context.Context, body VoteRequest) error {
-	log := logger.FromContext(ctx)
-	err := s.database.RemoveVote(body.Board, body.User, body.Note)
-	if err != nil {
-		log.Errorw("unable to remove vote", "board", body.Board, "user", body.User)
-	}
-	return err
-}
-
-func (s Service) GetVotes(ctx context.Context, f filter.VoteFilter) ([]*Vote, error) {
-	log := logger.FromContext(ctx)
-	votes, err := s.database.GetVotes(f)
-	if err != nil {
-		log.Errorw("unable to get votes", "err", err)
-	}
-	return Votes(votes), err
-}
-
-func (s *Service) getVotes(ctx context.Context, boardID, id uuid.UUID) ([]VoteDB, error) {
-	log := logger.FromContext(ctx)
-	votes, err := s.database.GetVotes(filter.VoteFilter{Board: boardID, Voting: &id})
-	if err != nil {
-		log.Errorw("unable to get votes", "voting", id, "error", err)
-		return nil, err
-	}
-	return votes, err
-}
-
-func (s *Service) createdVoting(board, voting uuid.UUID) {
-	dbVoting, _, err := s.database.GetVoting(board, voting)
-	if err != nil {
-		logger.Get().Errorw("unable to get voting in created voting", "err", err)
-		return
+func getVotingWithResults(voting database.Voting, votes []database.Vote) *VotingResults {
+	if voting.Status != types.VotingStatusClosed {
+		return nil
 	}
 
-	_ = s.realtime.BroadcastToBoard(board, realtime.BoardEvent{
-		Type: realtime.BoardEventVotingCreated,
-		Data: new(Voting).From(dbVoting, nil),
+	relevantVoting := technical_helper.Filter[database.Vote](votes, func(vote database.Vote) bool {
+		return vote.Voting == voting.ID
 	})
-}
 
-func (s *Service) updatedVoting(board, voting uuid.UUID) {
-	//todo: fix notes
-	var notes []database.Note
-	dbVoting, dbVotes, err := s.database.GetVoting(board, voting)
-	if err != nil {
-		logger.Get().Errorw("unable to retrieve voting in updated voting", "err", err)
-		return
+	if len(relevantVoting) <= 0 {
+		return nil
 	}
-	if dbVoting.Status == types.VotingStatusClosed {
-		//notes, err = s.database.GetNotes(board)
-		if err != nil {
-			logger.Get().Errorw("unable to retrieve notes in updated voting", "err", err)
+
+	votingResult := VotingResults{Total: len(relevantVoting), Votes: map[uuid.UUID]VotingResultsPerNote{}}
+	totalVotePerNote := map[uuid.UUID]int{}
+	votesPerUser := map[uuid.UUID][]uuid.UUID{}
+	for _, vote := range relevantVoting {
+		if _, ok := totalVotePerNote[vote.Note]; ok {
+			totalVotePerNote[vote.Note] = totalVotePerNote[vote.Note] + 1
+			votesPerUser[vote.Note] = append(votesPerUser[vote.Note], vote.User)
+		} else {
+			totalVotePerNote[vote.Note] = 1
+			votesPerUser[vote.Note] = []uuid.UUID{vote.User}
 		}
 	}
 
-	_ = s.realtime.BroadcastToBoard(board, realtime.BoardEvent{
-		Type: realtime.BoardEventVotingUpdated,
-		Data: struct {
-			Voting *Voting        `json:"voting"`
-			Notes  []*notes2.Note `json:"notes"`
-		}{
-			Voting: new(Voting).From(dbVoting, dbVotes),
-			Notes:  notes2.Notes(notes),
-		},
-	})
+	for note, total := range totalVotePerNote {
+		result := VotingResultsPerNote{
+			Total: total,
+		}
+		if voting.ShowVotesOfOthers {
+			userVotes := map[uuid.UUID]int{}
+			for _, user := range votesPerUser[note] {
+				if _, ok := userVotes[user]; ok {
+					userVotes[user] = userVotes[user] + 1
+				} else {
+					userVotes[user] = 1
+				}
+			}
 
+			var votingResultsPerUser []VotingResultsPerUser
+			for user, total := range userVotes {
+				votingResultsPerUser = append(votingResultsPerUser, VotingResultsPerUser{
+					ID:    user,
+					Total: total,
+				})
+			}
+
+			result.Users = &votingResultsPerUser
+		}
+
+		votingResult.Votes[note] = result
+	}
+	return &votingResult
 }
 
-func NewVotingService(db *VotingDatabase, rt *realtime.Broker) *Service {
-	b := new(Service)
-	b.database = *db
-	b.realtime = rt
-	return b
+func (v *Voting) calculateTotalVoteCount(notes notes.NoteSlice) *VotingResults {
+	totalVotingCount := 0
+	votingResultsPerNode := &VotingResults{
+		Votes: make(map[uuid.UUID]VotingResultsPerNote),
+	}
+
+	for _, note := range notes {
+		if voteResults, ok := v.VotingResults.Votes[note.ID]; ok { // Check if note was voted on
+			votingResultsPerNode.Votes[note.ID] = VotingResultsPerNote{
+				Total: voteResults.Total,
+				Users: voteResults.Users,
+			}
+			totalVotingCount += v.VotingResults.Votes[note.ID].Total
+		}
+	}
+
+	votingResultsPerNode.Total = totalVotingCount
+
+	return votingResultsPerNode
+}
+
+func (v *Voting) hasNoResults() bool {
+	return v.VotingResults == nil
 }
