--- server/src/votes/votings_test.go
@@ -1,1 +1,0 @@
-package votes
